// packages/adi-core/src/documents/models/CaseDocument.ts
import { DocumentSource } from "../enums/DocumentSource";
import { DocumentStatus } from "../enums/DocumentStatus";
import { DocumentType } from "../enums/DocumentType";

export interface CaseDocument {
  id: string;

  type: DocumentType;
  status: DocumentStatus;
  source: DocumentSource;

  originalFileName?: string;
  mimeType?: string;
  pageCount?: number;

  confidence?: number;
  rawText?: string;
  parsedData?: unknown;

  createdAt: Date;
  updatedAt?: Date;
}